refactor: extract slider TypeScript types into dedicated file (Sat, May 31, 2025 12:05 AM)
