+++ src/types/slider.ts
@@ -1,0 +1,12 @@
+interface ImageItem {
+  url: string;
+}
+
+export interface SliderDetailProps {
+  id: string;
+  name: {
+    en?: string;
+    ru?: string;
+  };
+  images: ImageItem[];
+}
